create util mod and move smoothjoin to merge
