+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod merge;
+pub mod util;
